src/backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
boto3
langchain
chromadb
langchain-aws
python-dotenv
langgraph
